V 1.9 deposit & withdraw added

+++ src/interfaces/Payable.java
@@ -1,6 +1,6 @@
 package interfaces;
 
 public interface Payable {
-    void pay(double amount);
+    void pay(int amount);
     double getDueAmount();
 }
